:truck: entity -> repository 패키지로 이동
