+++ src/test/java/orgo/backend/domain/_5badge/repository/RecordHeightBadgeRepositoryTest.java
@@ -1,4 +1,4 @@
-package orgo.backend.domain._5badge.entity;
+package orgo.backend.domain._5badge.repository;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
@@ -6,6 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import orgo.backend.domain._3mountain.entity.Mountain;
 import orgo.backend.domain._3mountain.repository.MountainRepository;
+import orgo.backend.domain._5badge.entity.Badge;
+import orgo.backend.domain._5badge.entity.BadgeGroup;
+import orgo.backend.domain._5badge.entity.RecordHeightBadge;
 import orgo.backend.domain._5badge.repository.BadgeRepository;
 import orgo.backend.setting.MockEntityFactory;
 import orgo.backend.setting.RepositoryTest;
@@ -15,7 +18,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
-public class RecordCountBadgeRepositoryTest extends RepositoryTest {
+public class RecordHeightBadgeRepositoryTest extends RepositoryTest {
 
     @Autowired
     BadgeRepository<Badge> badgeRepository;
@@ -24,19 +27,19 @@ public class RecordCountBadgeRepositoryTest extends RepositoryTest {
     MountainRepository mountainRepository;
 
     @Test
-    @DisplayName("RecordCountBadge를 저장하고, 조회한다.")
+    @DisplayName("RecordHeightBadge를 저장하고, 조회한다.")
     void saveAndFind() {
         //given
         Mountain mountain = mountainRepository.save(MockEntityFactory.mockMountain(null, MockEntityFactory.mockPeak(null)));
-        RecordCountBadge recordCountBadge = RecordCountBadge.builder()
+        RecordHeightBadge recordCountBadge = RecordHeightBadge.builder()
                 .mainGroup(BadgeGroup.RECORD)
                 .mountain(mountain)
-                .count(1)
+                .height(100.5)
                 .build();
 
         //when
-        RecordCountBadge saved = badgeRepository.save(recordCountBadge);
-        RecordCountBadge found = (RecordCountBadge) badgeRepository.findById(saved.getId()).get();
+        RecordHeightBadge saved = badgeRepository.save(recordCountBadge);
+        RecordHeightBadge found = (RecordHeightBadge) badgeRepository.findById(saved.getId()).get();
         log.info("{}", found);
 
         //then
