+++ src/test/java/orgo/backend/domain/_5badge/repository/RecordMonthBadgeRepositoryTest.java
@@ -1,4 +1,4 @@
-package orgo.backend.domain._5badge.entity;
+package orgo.backend.domain._5badge.repository;
 
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.DisplayName;
@@ -6,16 +6,19 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import orgo.backend.domain._3mountain.entity.Mountain;
 import orgo.backend.domain._3mountain.repository.MountainRepository;
+import orgo.backend.domain._5badge.entity.Badge;
+import orgo.backend.domain._5badge.entity.BadgeGroup;
+import orgo.backend.domain._5badge.entity.RecordMonthBadge;
 import orgo.backend.domain._5badge.repository.BadgeRepository;
 import orgo.backend.setting.MockEntityFactory;
 import orgo.backend.setting.RepositoryTest;
 
-import java.util.List;
+import java.time.YearMonth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Slf4j
-public class RecordHeightBadgeRepositoryTest extends RepositoryTest {
+public class RecordMonthBadgeRepositoryTest extends RepositoryTest {
 
     @Autowired
     BadgeRepository<Badge> badgeRepository;
@@ -24,19 +27,19 @@ public class RecordHeightBadgeRepositoryTest extends RepositoryTest {
     MountainRepository mountainRepository;
 
     @Test
-    @DisplayName("RecordHeightBadge를 저장하고, 조회한다.")
+    @DisplayName("RecordMonthBadge를 저장하고, 조회한다.")
     void saveAndFind() {
         //given
         Mountain mountain = mountainRepository.save(MockEntityFactory.mockMountain(null, MockEntityFactory.mockPeak(null)));
-        RecordHeightBadge recordCountBadge = RecordHeightBadge.builder()
+        RecordMonthBadge recordCountBadge = RecordMonthBadge.builder()
                 .mainGroup(BadgeGroup.RECORD)
                 .mountain(mountain)
-                .height(100.5)
+                .yearMonth(YearMonth.of(2023, 9))
                 .build();
 
         //when
-        RecordHeightBadge saved = badgeRepository.save(recordCountBadge);
-        RecordHeightBadge found = (RecordHeightBadge) badgeRepository.findById(saved.getId()).get();
+        RecordMonthBadge saved = badgeRepository.save(recordCountBadge);
+        RecordMonthBadge found = (RecordMonthBadge) badgeRepository.findById(saved.getId()).get();
         log.info("{}", found);
 
         //then
